Autodetect path for assemble and run scripts
Odo hardcodes the S2I assemble script path in
`https://github.com/redhat-developer/odo-supervisord-image/blob/master/assemble-and-restart`
But the S2I docs suggest that the S2I scripts need to looked up in the
following order:
1. A script found at the --scripts-url URL
2. A script found in the application source .s2i/bin directory
3. A script found at the default image URL (io.openshift.s2i.scripts-url label)
and ultimately giveup if not found in any of the above paths.

While, 1 is not relevant to us(builds triggered internally by odo), 2 is already
done by odo, 3 is enhanced by this PR instead of using the currently(without this PR)
hard-coded path

This PR uses the env vars passed by odo due to
`https://github.com/redhat-developer/odo/pull/924` for running assemble script.

Note: The S2I scripts suggest that the builder image might point to any of the following
paths:

a. image://path_to_scripts_dir - absolute path inside the image
b. file://path_to_scripts_dir - relative or absolute path on the host machine
c. http(s)://path_to_scripts_dir - URL to a directory

Of these, the assemble-and-restart script is with this PR currently capable of handling
`image://path_to_scripts_dir` but not `file://` or `http(s)://` and they are expected to be
added over time. At the moment for any of the `file://` or `http(s)://`, the component
create and/or update is expected to fail as it is expected to be even with latest master.

To test:
Please refer https://github.com/redhat-developer/odo/pull/924 steps to test

fixes: redhat-developer/odo#869
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>